refactor(filter): 필터 내에 Response값 반환 형태로 변경
데이터의 불변성을 유지하기 위해서 필터 내에 Response 값을 반환인자로 넣어줬음.
